Allow building on non-linux
