+++ pkg/peertracker/trace_others.go
@@ -1,0 +1,13 @@
+//go:build !linux
+// +build !linux
+
+package peertracker
+
+import (
+	"errors"
+	"net"
+)
+
+func CallerFromVSockConn(conn net.Conn) (CallerInfo, error) {
+	return CallerInfo{}, errors.New("not supported")
+}
